PeakFinder algorithm is implemented and tested
