+++ src/tests/algorithms_test/standard_deviation_test.cpp
@@ -8,43 +8,40 @@ using namespace Catch;
 TEST_CASE("StandardDeviation add and stddev calculation", "[StandardDeviation]") {
 StandardDeviation stddev;
 
-REQUIRE_THROWS_AS(stddev.stddev(), std::runtime_error);
+    REQUIRE(stddev.stddev() == Approx(0));
 
-stddev.add(1.0);
-REQUIRE_THROWS_AS(stddev.stddev(), std::runtime_error);
+    stddev.add(1.0);
+    REQUIRE(stddev.stddev() == Approx(0));
 
-stddev.add(2.0);
-REQUIRE(stddev.stddev() == Approx(0.707106781));
+    stddev.add(2.0);
+    REQUIRE(stddev.stddev() == Approx(0.5));
 
-stddev.add(3.0);
-REQUIRE(stddev.stddev() == Approx(1.0));
+    stddev.add(3.0);
+    REQUIRE(stddev.stddev() == Approx(0.8164965809));
 
-stddev.add(4.0);
-REQUIRE(stddev.stddev() == Approx(1.290994449));
+    stddev.add(4.0);
+    REQUIRE(stddev.stddev() == Approx(1.1180339887499));
 }
 
 TEST_CASE("StandardDeviation handles negative numbers", "[StandardDeviation]") {
-StandardDeviation stddev;
+    StandardDeviation stddev;
 
-stddev.add(-1.0);
-stddev.add(2.0);
-REQUIRE(stddev.stddev() == Approx(2.121320344));
+    stddev.add(-1.0);
+    stddev.add(2.0);
+    REQUIRE(stddev.stddev() == Approx(1.5));
 
-stddev.add(-3.0);
-stddev.add(4.0);
-REQUIRE(stddev.stddev() == Approx(2.828427125));
+    stddev.add(-3.0);
+    stddev.add(4.0);
+    REQUIRE(stddev.stddev() == Approx(2.6925824035673));
 }
 
 TEST_CASE("StandardDeviation handles large numbers", "[StandardDeviation]") {
-StandardDeviation stddev;
+    StandardDeviation stddev;
 
-stddev.add(1e12);
-stddev.add(2e12);
-stddev.add(3e12);
-REQUIRE(stddev.stddev() == Approx(1e12));
-
-stddev.add(4e12);
-REQUIRE(stddev.stddev() == Approx(1.290994449e12));
+    stddev.add(1e12);
+    stddev.add(2e12);
+    stddev.add(3e12);
+    REQUIRE(stddev.stddev() == Approx(0.81649658e12));
 }
 
 TEST_CASE("StandardDeviation implicit conversion to double", "[StandardDeviation]") {
@@ -55,5 +52,5 @@ stddev.add(2.0);
 stddev.add(3.0);
 
 double sd = stddev;
-REQUIRE(sd == Approx(1.0));
+REQUIRE(sd == Approx(0.81649658092773));
 }
